Exception handling - delete

+++ src/main/java/com/programa/programa/services/UserServices.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.programa.programa.entities.User;
 import com.programa.programa.repositories.UserRepository;
+import com.programa.programa.services.exceptions.DatabaseException;
 import com.programa.programa.services.exceptions.ResourceNotFoundException;
 
 @Service //component
@@ -30,8 +32,16 @@ public class UserServices {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);;
-	}
+		try {
+	        if (repository.existsById(id)) {
+	            repository.deleteById(id);			
+	        } else {				
+	            throw new ResourceNotFoundException(id);			
+	        }		
+	    } catch (DataIntegrityViolationException e) {			
+	        throw new DatabaseException(e.getMessage());		
+	    }	
+	}	
 	
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
